+++ Makefile.toml
@@ -12,9 +12,13 @@ args = ["clean"]
 command = "rustup"
 args = ["add", "target", "wasm32-unknown-unknown"]
 
+[tasks.check]
+command = "cargo"
+args = ["check", "--all-features"]
+
 [tasks.build]
 command = "cargo"
-args = ["build", "--release", "--all-features"]
+args = ["build", "--release"]
 
 [tasks.test]
 command = "cargo"
@@ -108,7 +112,7 @@ dependencies = ["format", "zepter"]
 dependencies = ["audit", "machete"]
 
 [tasks.ci-common]
-dependencies = ["build", "test", "clippy", "dependencies-check"]
+dependencies = ["check", "build", "test", "clippy", "dependencies-check"]
 
 [tasks.ci]
 dependencies = ["quick-fix", "ci-common"]
