--- runtime/src/configs/aura.rs
@@ -1,30 +1,0 @@
-// Copyright 2025, Horizen Labs, Inc.
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-use frame_support::traits::ConstU64;
-
-use crate::{
-    configs::{AllowMultipleBlocksPerSlot, MaxAuthorities},
-    constants::SLOT_DURATION,
-    AuraId, Runtime,
-};
-
-impl pallet_aura::Config for Runtime {
-    type AllowMultipleBlocksPerSlot = AllowMultipleBlocksPerSlot;
-    type AuthorityId = AuraId;
-    type DisabledValidators = ();
-    type MaxAuthorities = MaxAuthorities;
-    type SlotDuration = ConstU64<SLOT_DURATION>;
-}
