+++ runtime/src/configs/consensus.rs
@@ -13,30 +13,39 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
+//! In this module, we provide the configurations about parachain consensus.
+
 use crate::{
-    constants::HOURS, weights, AccountId, Balances, BlockNumber, CollatorSelection, Runtime,
+    constants::{HOURS, SLOT_DURATION},
+    weights, AccountId, Aura, AuraId, Balances, BlockNumber, CollatorSelection, Runtime,
     RuntimeEvent, Session, SessionKeys,
 };
-use frame_support::{parameter_types, traits::ConstU32, PalletId};
+use frame_support::{parameter_types, traits::ConstU64, PalletId};
 use frame_system::EnsureRoot;
+
+impl pallet_authorship::Config for Runtime {
+    type FindAuthor = pallet_session::FindAccountFromAuthorIndex<Self, Aura>;
+    type EventHandler = ();
+}
+
+pub type CollatorSelectionUpdateOrigin = EnsureRoot<AccountId>;
 
 parameter_types! {
     pub const PotId: PalletId = PalletId(*b"PotStake");
     pub const SessionLength: BlockNumber = 6 * HOURS;
-    // // StakingAdmin pluralistic body.
-    // pub const StakingAdminBodyId: BodyId = BodyId::Defense;
+    pub const MaxCandidates: u32 = 30;
+    pub const MaxInvulnerables: u32 = 10;
+    pub const MinEligibleCollators: u32 = 1;
 }
-
-pub type CollatorSelectionUpdateOrigin = EnsureRoot<AccountId>;
 
 impl pallet_collator_selection::Config for Runtime {
     type RuntimeEvent = RuntimeEvent;
     type Currency = Balances;
     type UpdateOrigin = CollatorSelectionUpdateOrigin;
     type PotId = PotId;
-    type MaxCandidates = ConstU32<50>;
-    type MinEligibleCollators = ConstU32<2>;
-    type MaxInvulnerables = ConstU32<5>;
+    type MaxCandidates = MaxCandidates;
+    type MinEligibleCollators = MinEligibleCollators;
+    type MaxInvulnerables = MaxInvulnerables;
     // should be a multiple of session or things will get inconsistent
     type KickThreshold = Period;
     type ValidatorId = <Self as frame_system::Config>::AccountId;
@@ -51,14 +60,29 @@ parameter_types! {
 }
 
 impl pallet_session::Config for Runtime {
-    type Keys = SessionKeys;
-    type NextSessionRotation = pallet_session::PeriodicSessions<Period, Offset>;
     type RuntimeEvent = RuntimeEvent;
-    // Essentially just Aura, but let's be pedantic.
-    type SessionHandler = <SessionKeys as sp_runtime::traits::OpaqueKeys>::KeyTypeIdProviders;
-    type SessionManager = CollatorSelection;
-    type ShouldEndSession = pallet_session::PeriodicSessions<Period, Offset>;
     type ValidatorId = AccountId;
     type ValidatorIdOf = pallet_collator_selection::IdentityCollator;
+    type ShouldEndSession = pallet_session::PeriodicSessions<Period, Offset>;
+    type NextSessionRotation = pallet_session::PeriodicSessions<Period, Offset>;
+    type SessionManager = CollatorSelection;
+    // Essentially just Aura, but let's be pedantic.
+    type SessionHandler = <SessionKeys as sp_runtime::traits::OpaqueKeys>::KeyTypeIdProviders;
+    type Keys = SessionKeys;
     type WeightInfo = weights::pallet_session::ZKVEvmWeight<Runtime>;
 }
+
+parameter_types! {
+    pub const AllowMultipleBlocksPerSlot: bool = true;
+    pub const MaxAuthorities: u32 = 100_000;
+}
+
+impl pallet_aura::Config for Runtime {
+    type AuthorityId = AuraId;
+    type MaxAuthorities = MaxAuthorities;
+    type DisabledValidators = ();
+    type AllowMultipleBlocksPerSlot = AllowMultipleBlocksPerSlot;
+    type SlotDuration = ConstU64<SLOT_DURATION>;
+}
+
+impl cumulus_pallet_aura_ext::Config for Runtime {}
