--- runtime/src/configs/timestamp.rs
@@ -1,26 +1,0 @@
-// Copyright 2025, Horizen Labs, Inc.
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-use frame_support::traits::ConstU64;
-
-use crate::{constants::SLOT_DURATION, weights, Aura, Runtime};
-
-impl pallet_timestamp::Config for Runtime {
-    type MinimumPeriod = ConstU64<{ SLOT_DURATION / 2 }>;
-    /// A timestamp: milliseconds since the unix epoch.
-    type Moment = u64;
-    type OnTimestampSet = Aura;
-    type WeightInfo = weights::pallet_timestamp::ZKVEvmWeight<Runtime>;
-}
