--- runtime/src/configs/transaction_payment.rs
@@ -1,36 +1,0 @@
-// Copyright 2025, Horizen Labs, Inc.
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-use frame_support::{parameter_types, weights::ConstantMultiplier};
-use polkadot_runtime_common::SlowAdjustingFeeUpdate;
-
-use crate::{
-    constants::currency::MICROCENTS, Balance, Balances, Runtime, RuntimeEvent, WeightToFee,
-};
-
-parameter_types! {
-    /// Relay Chain `TransactionByteFee` / 10
-    pub const TransactionByteFee: Balance = 10 * MICROCENTS;
-    pub const OperationalFeeMultiplier: u8 = 5;
-}
-
-impl pallet_transaction_payment::Config for Runtime {
-    type FeeMultiplierUpdate = SlowAdjustingFeeUpdate<Self>;
-    type LengthToFee = ConstantMultiplier<Balance, TransactionByteFee>;
-    type OnChargeTransaction = pallet_transaction_payment::FungibleAdapter<Balances, ()>;
-    type OperationalFeeMultiplier = OperationalFeeMultiplier;
-    type RuntimeEvent = RuntimeEvent;
-    type WeightToFee = WeightToFee;
-}
